home

+++ src/RestTraining/Controllers/ClientsController.cs
@@ -1,4 +1,7 @@
+using System;
 using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
 using System.Web.Http;
 using RestTraining.Domain;
 using RestTraining.Api.Models;
@@ -23,30 +26,55 @@ namespace RestTraining.Api.Controllers
         // GET api/Clients/5
         public Client Get(int id)
         {
+            var client = _clientRepository.Find(id);
+            if (client == null)
+            {
+                throw new HttpResponseException(new HttpResponseMessage(HttpStatusCode.NotFound));
+            } 
             return _clientRepository.Find(id);
         }
 
         // DELETE api/values/5
         public void Delete(int id)
         {
+            var client = _clientRepository.Find(id);
+            if (client == null)
+            {
+                throw new HttpResponseException(new HttpResponseMessage(HttpStatusCode.NotFound));
+            } 
             _clientRepository.Delete(id);
             _clientRepository.Save();
         }
         
         // POST api/values
-        public void Post(Client client)
+        //public HttpResponseMessage Post([ModelBinder(typeof(CustomPersonModelBinderProvider))]Client client)
+        public HttpResponseMessage Post(Client client)
         {
-            //Add
+            if (!ModelState.IsValid)
+            {
+                throw new HttpResponseException(new HttpResponseMessage(HttpStatusCode.BadRequest));
+            }
             _clientRepository.InsertOrUpdate(client);
             _clientRepository.Save();
+            var response = Request.CreateResponse(HttpStatusCode.Created, client);
+            string uri = Url.Route(null, new { id = client.Id });
+            response.Headers.Location = new Uri(Request.RequestUri, uri);
+            return response;
         }
 
         // PUT api/Clients/5
-        public void Put(Client client)
+        public HttpResponseMessage Put(Client client)
         {
-            //Update
+            if (!ModelState.IsValid)
+            {
+                throw new HttpResponseException(new HttpResponseMessage(HttpStatusCode.BadRequest));
+            }
             _clientRepository.InsertOrUpdate(client);
             _clientRepository.Save();
+            var response = Request.CreateResponse(HttpStatusCode.OK, client);
+            string uri = Url.Route(null, new { id = client.Id });
+            response.Headers.Location = new Uri(Request.RequestUri, uri);
+            return response;
         }
 
         protected override void Dispose(bool disposing)
